Change standalone sync script into a Django custom management command
