+++ docs/quickStart.md
@@ -74,5 +74,5 @@ If you're working with the AdOps dashboard, you may want to pull in Boostr Deals
 
 ``` shell
 docker exec -it consvc-shepherd-app-1 sh # interactive mode
-BOOSTR_BASE_URL=https://app.boostr.com/api/ BOOSTR_EMAIL=find-me-in-1pass-ads-eng-vault@mozilla.com BOOSTR_PASSWORD=secret-from-1pass python sync_boostr_data.py
+python manage.py sync_boostr_data https://app.boostr.com/api/ find-me-in-1pass-ads-eng-vault@mozilla.com secret-from-1pass
 ```
